+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+
+- Telemetry events.
+
+### Fixed
+
+- Decrement counter when max retries has been reached.
+
 ## [0.1.0] - 2020-05-16
 
 Initial release.